Create HTML table from CSV or JSON
A script to take a JSON or CSV data feed and return it as a formatted
table.
Use the CSV option when you need the items in the order they're
presented in the CSV.
The script was written for Socrata data but should work for any JSON or
CSV.
The CSS classes are Twitter Bootstrap.